refactor: move renderer to new module
