+++ crates/render/src/renderer.rs
@@ -1,0 +1,187 @@
+use anyhow::{Context, Result};
+use physics::Sphere;
+use wgpu::util::DeviceExt;
+use winit::event::{Event, WindowEvent};
+use std::time::Duration;
+use winit::event_loop::{EventLoop};
+use winit::window::WindowBuilder;
+use winit::platform::pump_events::{EventLoopExtPumpEvents, PumpStatus};
+
+pub struct Renderer<'w> {
+    event_loop: EventLoop<()>,
+    window: winit::window::Window,
+    surface: wgpu::Surface<'w>,
+    device: wgpu::Device,
+    queue: wgpu::Queue,
+    config: wgpu::SurfaceConfiguration,
+    pipeline: wgpu::RenderPipeline,
+    vertex_buffer: wgpu::Buffer,
+    vertices: Vec<[f32; 2]>,
+}
+
+impl<'w> Renderer<'w> {
+    pub fn new() -> Result<Renderer<'static>> {
+        let event_loop = EventLoop::new().context("create event loop")?;
+        let window = WindowBuilder::new()
+            .with_title("Arena Renderer")
+            .build(&event_loop)
+            .context("failed to create window")?;
+
+        let instance = wgpu::Instance::default();
+        let surface = instance.create_surface(&window)?;
+        let surface = unsafe { std::mem::transmute::<wgpu::Surface<'_>, wgpu::Surface<'static>>(surface) };
+        let adapter = pollster::block_on(instance.request_adapter(&wgpu::RequestAdapterOptions {
+            power_preference: wgpu::PowerPreference::HighPerformance,
+            compatible_surface: Some(&surface),
+            force_fallback_adapter: false,
+        }))
+        .context("failed to get adapter")?;
+
+        let (device, queue) = pollster::block_on(adapter.request_device(
+            &wgpu::DeviceDescriptor {
+                label: Some("Renderer Device"),
+                required_features: wgpu::Features::empty(),
+                required_limits: wgpu::Limits::default(),
+            },
+            None,
+        ))
+        .context("failed to request device")?;
+
+        let size = window.inner_size();
+        let formats = surface.get_capabilities(&adapter).formats;
+        let format = formats[0];
+        let config = wgpu::SurfaceConfiguration {
+            usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
+            format,
+            width: size.width,
+            height: size.height,
+            desired_maximum_frame_latency: 2,
+            present_mode: wgpu::PresentMode::Fifo,
+            alpha_mode: wgpu::CompositeAlphaMode::Auto,
+            view_formats: vec![],
+        };
+        surface.configure(&device, &config);
+
+        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
+            label: Some("point shader"),
+            source: wgpu::ShaderSource::Wgsl(include_str!("shader.wgsl").into()),
+        });
+
+        let pipeline_layout =
+            device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
+                label: Some("layout"),
+                bind_group_layouts: &[],
+                push_constant_ranges: &[],
+            });
+
+        let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+            label: Some("pipeline"),
+            layout: Some(&pipeline_layout),
+            vertex: wgpu::VertexState {
+                module: &shader,
+                entry_point: "vs_main",
+                buffers: &[wgpu::VertexBufferLayout {
+                    array_stride: std::mem::size_of::<[f32; 2]>() as u64,
+                    step_mode: wgpu::VertexStepMode::Vertex,
+                    attributes: &[wgpu::VertexAttribute {
+                        format: wgpu::VertexFormat::Float32x2,
+                        offset: 0,
+                        shader_location: 0,
+                    }],
+                }],
+            },
+            fragment: Some(wgpu::FragmentState {
+                module: &shader,
+                entry_point: "fs_main",
+                targets: &[Some(wgpu::ColorTargetState {
+                    format: config.format,
+                    blend: Some(wgpu::BlendState::REPLACE),
+                    write_mask: wgpu::ColorWrites::ALL,
+                })],
+            }),
+            primitive: wgpu::PrimitiveState {
+                topology: wgpu::PrimitiveTopology::PointList,
+                ..Default::default()
+            },
+            depth_stencil: None,
+            multisample: wgpu::MultisampleState::default(),
+            multiview: None,
+        });
+
+        let vertices: Vec<[f32; 2]> = Vec::new();
+        let vertex_buffer = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
+            label: Some("vertices"),
+            contents: bytemuck::cast_slice(&vertices),
+            usage: wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST,
+        });
+
+        Ok(Renderer {
+            event_loop,
+            window,
+            surface,
+            device,
+            queue,
+            config,
+            pipeline,
+            vertex_buffer,
+            vertices,
+        })
+    }
+
+    pub fn update_spheres(&mut self, spheres: &[Sphere]) {
+        self.vertices.clear();
+        for s in spheres {
+            self.vertices.push([s.pos.x * 0.1, -s.pos.y * 0.1]);
+        }
+        if !self.vertices.is_empty() {
+            self.queue
+                .write_buffer(&self.vertex_buffer, 0, bytemuck::cast_slice(&self.vertices));
+        }
+    }
+
+    pub fn render(&mut self) -> Result<()> {
+        let status = self.event_loop.pump_events(Some(Duration::ZERO), |event, elwt| {
+            if let Event::WindowEvent { event: WindowEvent::CloseRequested, .. } = &event {
+                elwt.exit();
+            }
+        });
+        if matches!(status, PumpStatus::Exit(_)) {
+            return Ok(());
+        }
+
+        let output = self
+            .surface
+            .get_current_texture()
+            .context("failed to acquire surface texture")?;
+        let view = output
+            .texture
+            .create_view(&wgpu::TextureViewDescriptor::default());
+
+        let mut encoder =
+            self.device
+                .create_command_encoder(&wgpu::CommandEncoderDescriptor { label: Some("enc") });
+        {
+            let mut rpass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
+                label: Some("rpass"),
+                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
+                    view: &view,
+                    resolve_target: None,
+                    ops: wgpu::Operations {
+                        load: wgpu::LoadOp::Clear(wgpu::Color::BLACK),
+                        store: wgpu::StoreOp::Store,
+                    },
+                })],
+                depth_stencil_attachment: None,
+                timestamp_writes: None,
+                occlusion_query_set: None,
+            });
+            rpass.set_pipeline(&self.pipeline);
+            rpass.set_vertex_buffer(0, self.vertex_buffer.slice(..));
+            rpass.draw(0..self.vertices.len() as u32, 0..1);
+        }
+        self.queue.submit(Some(encoder.finish()));
+        output.present();
+        Ok(())
+    }
+}
+
